requirements.txt
fastapi==0.115.12
pydantic==2.11.5
python-dotenv==1.1.0
rapidfuzz==3.13.0
starlette==0.47.0
